mdp4_overlay_lcdc: Fix wrong initialization of VSYNC
COMMIT:msm: mdp: Initialize vsync sysfs in probe
SHA: 41db433cc845eb27bb765cc0cf599e872edd709f

The commit reported above change the initialization of the vsync
and sets the vsync in suspend mode when you boot.

Maybe it wanted but in our device,
cause the interruption of the VSYNC in the boot,
by creating a black screen.